Убрал метод сортировки в файл tasl_sort.service.dart
